backend/requirements.txt
asgiref==3.11.0
boto3==1.42.12
botocore==1.42.12
diff-match-patch==20241021
Django==4.2.27
django-cors-headers==4.9.0
django-import-export==4.3.14
django-storages==1.14.6
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
jmespath==1.0.1
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
tablib==3.9.0
typing_extensions==4.15.0
urllib3==1.26.20
xlrd==2.0.2
xlwt==1.3.0
